+++ lib/rubocop/cop/rspec_file_name.rb
@@ -10,45 +10,32 @@ module Rubocop
     class RSpecFileName < Cop
       include TopLevelDescribe
 
-      CLASS_SPEC_MSG = 'Class unit spec should have a path ending with `%s`'
-      METHOD_SPEC_MSG = 'Unit spec should have a path matching `%s`'
+      MESSAGE = 'Spec path should end with `%s`'
       METHOD_STRING_MATCHER = /^[\#\.].+/
 
       def on_top_level_describe(node, args)
-        class_or_method = const_name(args.first)
-        return unless class_or_method
+        return unless single_top_level_describe?
+        object = const_name(args.first)
+        return unless object
 
-        path_parts = class_or_method.split('::').map do |part|
-          camel_to_underscore(part)
-        end
+        glob_matcher = matcher(object, args[1])
+        return if source_filename =~ regexp_from_glob(glob_matcher)
 
-        if !args[1]
-          check_class_spec(node, path_parts)
-        else
-          method_str = args[1].children.first if args[1]
-          path_parts << 'class_methods' if method_str.start_with? '.'
-          check_method_spec(node, path_parts, method_str)
-        end
+        add_offense(node, :expression, format(MESSAGE, glob_matcher))
       end
 
       private
 
-      def check_class_spec(node, path_parts)
-        spec_path = File.join(path_parts) + '_spec.rb'
-        return if source_filename.end_with? spec_path
-
-        add_offense(node, :expression, format(CLASS_SPEC_MSG, spec_path))
+      def matcher(object, method)
+        method_strg = method ? method.children.first.gsub(/\W+/, '') : nil
+        path = [File.join(path_parts(object)), method_strg].compact.join('_')
+        "#{path}*_spec.rb"
       end
 
-      def check_method_spec(node, path_parts, method_str)
-        matcher_parts = path_parts.dup
-        # Strip out symbols; it's not worth enforcing a vocabulary for them.
-        matcher_parts << method_str[1..-1].gsub(/\W+/, '*') + '_spec.rb'
-
-        glob_matcher = File.join(matcher_parts)
-        return if source_filename =~ regexp_from_glob(glob_matcher)
-
-        add_offense(node, :expression, format(METHOD_SPEC_MSG, glob_matcher))
+      def path_parts(object)
+        object.split('::').map do |part|
+          camel_to_underscore(part)
+        end
       end
 
       def source_filename