+++ spec/rubocop/cop/rspec_file_name_spec.rb
@@ -7,12 +7,22 @@ describe Rubocop::Cop::RSpecFileName do
 
   it 'checks the path' do
     inspect_source(cop,
-                   ["describe MyClass, '#foo' do; end"],
-                   'wrong_path/foo_spec.rb')
+                   ["describe MyClass, 'foo' do; end"],
+                   'wrong_path_foo_spec.rb')
     expect(cop.offenses.size).to eq(1)
     expect(cop.offenses.map(&:line).sort).to eq([1])
-    expect(cop.messages).to eq(['Unit spec should have a path matching ' \
-                                '`my_class/foo_spec.rb`'])
+    expect(cop.messages)
+      .to eq(['Spec path should end with `my_class_foo*_spec.rb`'])
+  end
+
+  it 'checks the path' do
+    inspect_source(cop,
+                   ["describe MyClass, '#foo' do; end"],
+                   'wrong_class_foo_spec.rb')
+    expect(cop.offenses.size).to eq(1)
+    expect(cop.offenses.map(&:line).sort).to eq([1])
+    expect(cop.messages)
+      .to eq(['Spec path should end with `my_class_foo*_spec.rb`'])
   end
 
   it 'checks class spec paths' do
@@ -21,13 +31,21 @@ describe Rubocop::Cop::RSpecFileName do
                    'wrong_class_spec.rb')
     expect(cop.offenses.size).to eq(1)
     expect(cop.offenses.map(&:line).sort).to eq([1])
-    expect(cop.messages).to eq(['Class unit spec should have a path ending ' \
-                                'with `my_class_spec.rb`'])
+    expect(cop.messages)
+      .to eq(['Spec path should end with `my_class*_spec.rb`'])
   end
 
   it 'skips specs that do not describe a class / method' do
     inspect_source(cop,
                    ["describe 'Test something' do; end"],
+                   'some/class_spec.rb')
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'skips specs that do have multiple top level describes' do
+    inspect_source(cop,
+                   ["describe MyClass, 'do_this' do; end",
+                    "describe MyClass, 'do_that' do; end"],
                    'some/class_spec.rb')
     expect(cop.offenses).to be_empty
   end
@@ -63,35 +81,42 @@ describe Rubocop::Cop::RSpecFileName do
   it 'checks instance methods' do
     inspect_source(cop,
                    ["describe Some::Class, '#inst' do; end"],
-                   'some/class/inst_spec.rb')
+                   'some/class_inst_spec.rb')
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'checks methods' do
+    inspect_source(cop,
+                   ["describe Some::Class, 'inst' do; end"],
+                   'some/class_inst_spec.rb')
     expect(cop.offenses).to be_empty
   end
 
   it 'checks class methods' do
     inspect_source(cop,
                    ["describe Some::Class, '.inst' do; end"],
-                   'some/class/class_methods/inst_spec.rb')
+                   'some/class_inst_spec.rb')
     expect(cop.offenses).to be_empty
   end
 
   it 'ignores non-alphanumeric characters' do
     inspect_source(cop,
                    ["describe Some::Class, '#pred?' do; end"],
-                   'some/class/pred_spec.rb')
+                   'some/class_pred_spec.rb')
     expect(cop.offenses).to be_empty
   end
 
   it 'allows flexibility with predicates' do
     inspect_source(cop,
                    ["describe Some::Class, '#thing?' do; end"],
-                   'some/class/thing_predicate_spec.rb')
+                   'some/class_thing_predicate_spec.rb')
     expect(cop.offenses).to be_empty
   end
 
   it 'allows flexibility with operators' do
     inspect_source(cop,
                    ["describe MyClass, '#<=>' do; end"],
-                   'my_class/spaceship_operator_spec.rb')
+                   'my_class_spaceship_operator_spec.rb')
     expect(cop.offenses).to be_empty
   end
 end